backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-anymail==9.1
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.9.0
django-import-export==3.2.0
django-unfo
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
humanize==4.6.0
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
tzdata==2023.3